fix: resolve compilation errors for build script
- Add missing resources (ic_glass, glass_blue, dimens)
- Fix audio imports and scope issues
- Fix camera imports and async usage
- Fix type mismatches and null safety
- Fix inline function visibility

+++ app/src/main/java/dev/synople/glassassistant/performance/OptimizedResultDisplay.kt
@@ -192,7 +192,7 @@ class OptimizedResultDisplay @JvmOverloads constructor(
 
         val lines = mutableListOf<String>()
         val words = text.split("\\s+".toRegex())
-        val availableWidth = width - (paddingPx * 2)
+        val availableWidth = (width - (paddingPx * 2)).toFloat()
 
         if (availableWidth <= 0) {
             // Fallback if view not measured yet
